chore: Extract ContextScope utility to doctest/parts/context_scope.h
