+++ doctest/parts/doctest_fwd.h
@@ -80,6 +80,7 @@ DOCTEST_MSVC_SUPPRESS_WARNING(4623) // default constructor was implicitly define
 #include <doctest/parts/public/test_case.h>
 #include <doctest/parts/public/decorators.h>
 #include <doctest/parts/public/exception_translator.h>
+#include <doctest/parts/public/context_scope.h>
 
 namespace doctest {
 
@@ -95,12 +96,6 @@ struct DOCTEST_INTERFACE MessageData
     assertType::Enum m_severity;
 };
 
-struct DOCTEST_INTERFACE IContextScope
-{
-    DOCTEST_DECLARE_INTERFACE(IContextScope)
-    virtual void stringify(std::ostream*) const = 0;
-};
-
 } // namespace doctest
 
 #include <doctest/parts/public/exceptions.h>
@@ -111,47 +106,6 @@ namespace doctest {
 namespace detail {
     template<typename T>
     int instantiationHelper(const T&) { return 0; }
-
-    // ContextScope base class used to allow implementing methods of ContextScope
-    // that don't depend on the template parameter in doctest.cpp.
-    struct DOCTEST_INTERFACE ContextScopeBase : public IContextScope {
-        ContextScopeBase(const ContextScopeBase&) = delete;
-
-        ContextScopeBase& operator=(const ContextScopeBase&) = delete;
-        ContextScopeBase& operator=(ContextScopeBase&&) = delete;
-
-        ~ContextScopeBase() override = default;
-
-    protected:
-        ContextScopeBase();
-        ContextScopeBase(ContextScopeBase&& other) noexcept;
-
-        void destroy();
-        bool need_to_destroy{true};
-    };
-
-    template <typename L> class ContextScope : public ContextScopeBase
-    {
-        L lambda_;
-
-    public:
-        explicit ContextScope(const L &lambda) : lambda_(lambda) {}
-        explicit ContextScope(L&& lambda) : lambda_(static_cast<L&&>(lambda)) { }
-
-        ContextScope(const ContextScope&) = delete;
-        ContextScope(ContextScope&&) noexcept = default;
-
-        ContextScope& operator=(const ContextScope&) = delete;
-        ContextScope& operator=(ContextScope&&) = delete;
-
-        void stringify(std::ostream* s) const override { lambda_(s); }
-
-        ~ContextScope() override {
-            if (need_to_destroy) {
-                destroy();
-            }
-        }
-    };
 
     struct DOCTEST_INTERFACE MessageBuilder : public MessageData
     {
@@ -191,11 +145,6 @@ DOCTEST_MSVC_SUPPRESS_WARNING_POP
         bool log();
         void react();
     };
-
-    template <typename L>
-    ContextScope<L> MakeContextScope(const L &lambda) {
-        return ContextScope<L>(lambda);
-    }
 } // namespace detail
 
 } // namespace doctest
